hpm apps release v1.11.1; update hpm_monitor

+++ middleware/hpm_monitor/core/monitor_task.h
@@ -10,7 +10,7 @@
 
 #include <stdint.h>
 
-typedef int (*monitor_channel_cb)(uint8_t *src, uint16_t length);
+typedef int (*monitor_channel_cb)(uint8_t *src, uint32_t length);
 
 void monitor_task_init(void);
